+++ lib/utils/getIconsAsComponent.js
