+++ lib/utils/index.js
@@ -1,5 +1,5 @@
-import getIconsAsComponent from './get-icon-component';
-import ghGet from './gh-get';
+import getIconsAsComponent from './getIconsAsComponent';
+import ghGet from './ghGet';
 import isset from './isset';
 
 export {
